Module Tezos_021_PsQuebec_test_helpers.Account

type t = {

pkh : Tezos_base.TzPervasives.Signature.Public_key_hash.t;

pk : Tezos_base.TzPervasives.Signature.Public_key.t;

sk : Tezos_base.TzPervasives.Signature.Secret_key.t;

}

type account = t

val known_accounts : t Tezos_base.TzPervasives.Signature.Public_key_hash.Table.t

val activator_account : account

val dummy_account : account

val new_account : ?rng_state:Stdlib.Random.State.t -> ?seed:Tezos_base.TzPervasives.Bytes.t -> ?algo:Tezos_base.TzPervasives.Signature.algo -> unit -> account

new_account ?rng_state ?seed ?algo () creates a new account with curve algo with the given seed (or rng_state to generate the seed) and add it to the global account state.

val add_account : t -> unit

val find : Tezos_base.TzPervasives.Signature.Public_key_hash.t -> t Tezos_base.TzPervasives.tzresult Lwt.t

val find_alternate : Tezos_base.TzPervasives.Signature.Public_key_hash.t -> t

val default_initial_balance : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t

4.000.000.000 tez

val generate_accounts : ?rng_state:Stdlib.Random.State.t -> int -> t list Tezos_base.TzPervasives.tzresult

generate_accounts ?rng_state n first frees the global account state then generates n random accounts with rng_state to generate the seed and adds them to the global account state.

val commitment_secret : Tezos_protocol_021_PsQuebec.Protocol.Blinded_public_key_hash.activation_code

val new_commitment : ?seed:Tezos_base.TzPervasives.Bytes.t -> unit -> (account * Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Commitment.t) Tezos_base.TzPervasives.tzresult Lwt.t

val pkh_of_contract_exn : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t -> Tezos_base.TzPervasives.Signature.Public_key_hash.t

Fails if the contract is not an implicit one

val make_bootstrap_account : ?balance:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t -> ?delegate_to:Tezos_base.TzPervasives.Signature.public_key_hash option -> ?consensus_key:Tezos_base.TzPervasives.Signature.public_key option -> t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Parameters.bootstrap_account

make_bootstrap_account ~initial_balance ~delegate_to account creates a Parameters.bootstrap_account from an account with the default or set values. default initial_balance is default_initial_balance, delegate_to is None and consensus_key is None.

val make_bootstrap_accounts : ?bootstrap_balances:int64 list -> ?bootstrap_delegations: Tezos_base.TzPervasives.Signature.public_key_hash option list -> ?bootstrap_consensus_keys: Tezos_base.TzPervasives.Signature.public_key option list -> t list -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Parameters.bootstrap_account list

make_bootstrap_accounts ~bootstrap_balances ~bootstrap_delegations ~bootstrap_consensus_keys accounts combines the lists accounts, bootstrap_balances, bootstrap_delegations and bootstrap_consensus_keys to create a list of Parameters.bootstrap_account using make_bootstrap_account.