Module Context.Delegate

type info = Tezos_protocol_plugin_021_PsQuebec.Delegate_services.info = {

full_balance : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t;

current_frozen_deposits : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t;

frozen_deposits : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t;

staking_balance : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t;

frozen_deposits_limit : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t option;

delegated_contracts : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t list;

delegated_balance : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t;

min_delegated_in_current_cycle : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t * Tezos_protocol_021_PsQuebec.Protocol.Level_repr.t option;

total_delegated_stake : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t;

staking_denominator : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Staking_pseudotoken.t;

deactivated : bool;

grace_period : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Cycle.t;

pending_denunciations : bool;

voting_info : Vote.delegate_info;

active_consensus_key : Tezos_protocol_021_PsQuebec.Environment.Signature.Public_key_hash.t;

pending_consensus_keys : (Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Cycle.t * Tezos_protocol_021_PsQuebec.Environment.Signature.Public_key_hash.t) list;

}

type stake = {

frozen : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t;

weighted_delegated : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t;

}

val info : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_plugin_021_PsQuebec.Delegate_services.info Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val full_balance : ?__LOC__:string -> t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

Calls RPCs /chains/<chain_id>/blocks/<blocsk_id>/contracts/<Implicit pkh>/full_balance and /chains/<chain_id>/blocks/<block_id>/delegates/<pkh>/own_full_balance, checks that both RPCs output the same value, then returns this value.

val current_frozen_deposits : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val initial_frozen_deposits : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

calls the RPC frozen_deposits: we're using a different name to be more easily distinguishable from current_frozen_deposits

val staking_balance : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val unstaked_frozen_deposits : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_plugin_021_PsQuebec.Plugin.Alpha_services.Delegate.deposit_per_cycle list Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val staking_denominator : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_021_PsQuebec.Environment.Z.t Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val frozen_deposits_limit : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t option Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val deactivated : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> bool Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val grace_period : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Cycle.t Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val voting_info : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Vote.delegate_info Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val consensus_key : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_plugin_021_PsQuebec.Delegate_services.consensus_keys_info Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val participation : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Delegate.For_RPC.participation_info Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val is_forbidden : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> bool Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t

val stake_for_cycle : t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Cycle.t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> stake Tezos_base.TzPervasives.tzresult Tezos_protocol_021_PsQuebec.Environment.Lwt.t