Module Tezos_021_PsQuebec_test_helpers.Cpmm_logic

module Simulate_raw : sig ... end

This is a simulation of the CPMM contract, as implemented in mligo in src/proto_alpha/lib_protocol/contracts/cpmm.mligo. The interested reader should look for comments in this file to gain a better understanding of the contract logic.

module Simulate : sig ... end