Module Tezos_021_PsQuebec_test_helpers.Expr

exception Expression_from_string

val from_string : ?check_micheline_indentation:bool -> string -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Script.expr

Parse a Michelson expression from string, raising an exception on error.

val toplevel_from_string : ?check_micheline_indentation:bool -> string -> Tezos_protocol_021_PsQuebec.Protocol.Michelson_v1_primitives.prim Tezos_micheline.Micheline.canonical

Parses a Michelson contract from string, raising an exception on error.

val to_string : Tezos_protocol_021_PsQuebec.Protocol.Script_repr.expr -> string