Module Sapling_helpers.Interpreter_helpers

include module type of struct include Common end

val memo_size_of_int : int -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Sapling.Memo_size.t

val int_of_memo_size : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Sapling.Memo_size.t -> int

val assert_true : (bool, 'a) Stdlib.result Lwt.t -> (unit, 'a) Stdlib.result Lwt.t

val assert_false : (bool, 'a) Stdlib.result Lwt.t -> (unit, 'a) Stdlib.result Lwt.t

val assert_some : ('a option, 'b) Stdlib.result Lwt.t -> ('a, 'b) Stdlib.result Lwt.t

val assert_none : ('a option, 'b) Stdlib.result Lwt.t -> (unit, 'b) Stdlib.result Lwt.t

val assert_error : ('a, 'b) Stdlib.result Lwt.t -> (unit, 'c) Stdlib.result Lwt.t

val print : ?prefix:string -> 'a Data_encoding.Encoding.t -> 'a -> unit

val to_hex : 'a -> 'a Data_encoding.Encoding.t -> string

val randomized_byte : ?pos:int -> 'a -> 'a Data_encoding.Encoding.t -> 'a

type wallet = Common.wallet = {

sk : Tezos_sapling.Core.Wallet.Spending_key.t;

vk : Tezos_sapling.Core.Wallet.Viewing_key.t;

}

val wallet_gen : unit -> wallet

val gen_addr : int -> Tezos_sapling.Core.Client.Viewing_key.t -> Tezos_sapling.Core.Client.Viewing_key.address list

val gen_nf : unit -> Tezos_sapling.Core.Client.Nullifier.t

val gen_cm_cipher : memo_size:int -> unit -> Tezos_sapling.Core.Client.Commitment.t * Tezos_sapling.Core.Client.Ciphertext.t

val client_state_of_diff : memo_size:int -> (Tezos_sapling.Core.Validator.Hash.t * Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Sapling.diff) -> Tezos_sapling.Storage.state

include module type of struct include Contract_helpers end

val init : ?hard_gas_limit_per_block: Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Gas.Arith.integral -> unit -> (Block.t * Tezos_protocol_environment_021_PsQuebec.Signature.Public_key_hash.t * Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t * Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

Initializes 2 addresses to do only operations plus one that will be used to bake.

val read_file : string -> string

Return contents of a given file as string.

val load_script : storage:string -> string -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Script.t

Loads a script from file.

val originate_contract : string -> string -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t -> Block.t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> (Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t * Block.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val fake_KT1 : Tezos_protocol_021_PsQuebec.Protocol.Contract_hash.t

val default_self : Tezos_protocol_021_PsQuebec.Protocol.Contract_hash.t

val default_payer : Tezos_base.TzPervasives.Signature.Public_key_hash.t

val default_sender : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t

val default_step_constants : Tezos_protocol_021_PsQuebec.Protocol.Script_interpreter.step_constants

val run_script : Tezos_raw_protocol_021_PsQuebec.Alpha_context.t -> ?logger:Tezos_raw_protocol_021_PsQuebec.Script_typed_ir.logger -> ?step_constants: Tezos_protocol_021_PsQuebec.Protocol.Script_interpreter.step_constants -> ?internal:bool -> string -> ?entrypoint:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Entrypoint.t -> storage:string -> parameter:string -> unit -> (Tezos_protocol_021_PsQuebec.Protocol.Script_interpreter.execution_result * Tezos_raw_protocol_021_PsQuebec.Alpha_context.context) Tezos_base.TzPervasives.tzresult Lwt.t

Helper function that parses and typechecks a script, its initial storage and parameters from strings. It then executes the typed script with the storage and parameters and returns the result.

The step_constants argument passes in some data which remains constant throughout script's execution, hence the name. This includes addresses of the sender and payer, the address of the smart contract, the amount of Tez transferred to it and so on.

An internal operation is an operation generated by smart contract's execution rather than by an implicit account.

val originate_contract_from_string_hash : script:string -> storage:string -> source_contract:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t -> baker:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Block.t -> (Tezos_protocol_021_PsQuebec.Protocol.Contract_hash.t * Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Script.t * Block.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val originate_contract_from_string : script:string -> storage:string -> source_contract:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t -> baker:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> Block.t -> (Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t * Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Script.t * Block.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val originate_contract_hash : string -> string -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t -> Block.t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> (Tezos_protocol_021_PsQuebec.Protocol.Contract_hash.t * Block.t * string, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

Returns a block in which the contract is originated. Also returns the associated anti-replay string and KT1 address.

val hex_shield : memo_size:int -> wallet -> string -> string

val transac_and_sync : memo_size:int -> Block.t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Script.lazy_expr -> int -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Contract.t -> Tezos_raw_protocol_021_PsQuebec.Contract_hash.t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.public_key_hash -> (Block.t * Tezos_sapling.Storage.state, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val shield : memo_size:int -> Tezos_sapling.Forge.Core.Spending_key.t -> int -> Tezos_sapling.Core.Wallet.Viewing_key.t -> (string -> 'a) -> string -> 'a list * int

val next_block : Block.t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Operation.packed -> (Block.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t