Module Slashing_helpers.Misbehaviour_repr

Helpers related to Protocol.Misbehaviour_repr.

val pp : Stdlib.Format.formatter -> Tezos_protocol_021_PsQuebec.Protocol.Misbehaviour_repr.t -> unit

val from_duplicate_operation : 'kind Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Kind.consensus Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.operation -> Tezos_protocol_021_PsQuebec.Protocol.Misbehaviour_repr.t

Builds a misbehaviour object from either of the duplicate (pre)attestations that constitute a double (pre)attestating event.

val check_from_duplicate_operation : loc:string -> Tezos_raw_protocol_021_PsQuebec.Misbehaviour_repr.t -> 'kind Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Kind.consensus Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.operation -> unit Tezos_base.TzPervasives.tzresult Lwt.t

check_from_duplicate_operation ~loc misbehaviour duplicate_op asserts that misbehaviour correctly describes a double signing event involving duplicate_op.

val from_duplicate_block : Block.t -> Tezos_protocol_021_PsQuebec.Protocol.Misbehaviour_repr.t Tezos_base.TzPervasives.tzresult

Builds a misbehaviour object from either of the duplicate blocks that constitute a double baking event.