Module Tezos_021_PsQuebec_test_helpers.Tez_helpers

type t = Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t

This module wraps the errors from the protocol and provides some helpful functions manipulating Tez.

val zero : t

val one_mutez : t

val one_cent : t

val fifty_cents : t

val one : t

val max_tez : t

val equal : t -> t -> bool

val (-?) : t -> t -> t Tezos_base.TzPervasives.tzresult

val sub_opt : t -> t -> t option

val (+?) : t -> t -> t Tezos_base.TzPervasives.tzresult

val (*?) : t -> int64 -> t Tezos_base.TzPervasives.tzresult

val (/?) : t -> int64 -> t Tezos_base.TzPervasives.tzresult

val (-!) : t -> t -> t

val (+!) : t -> t -> t

val (*!) : t -> int64 -> t

val (/!) : t -> int64 -> t

val to_mutez : t -> int64

val of_mutez : int64 -> t

val to_repr : t -> Tezos_protocol_021_PsQuebec.Protocol.Tez_repr.t

val of_repr : Tezos_protocol_021_PsQuebec.Protocol.Tez_repr.t -> t

val min : t -> t -> t

module Compare : Tezos_base.TzPervasives.Compare.S with type t := t

val pp : Stdlib.Format.formatter -> t -> unit

val to_string : t -> string

val ratio : t -> t -> Q.t

val mul_q : t -> Q.t -> Q.t

val of_int : int -> t

val of_q : round:[ `Down | `Up ] -> Q.t -> t

val of_z : Z.t -> t

val to_z : t -> Z.t

module Ez_tez : sig ... end

Functions to manipulate Tez in a high level way