Module Tezos_021_PsQuebec_test_helpers.Tez_staking_helpers

Tez_staking_helpers defines different kinds of tez Modules that manipulate them in different ways. They involve more complicated operations, as they are related to staking, thus are represented as partial amounts, and are related to pseudotokens.

module Cycle = Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Cycle

module Tez : sig ... end

module Partial_tez : sig ... end

Representation of Tez with non integer values

module Frozen_tez : sig ... end

Frozen_tez represents frozen stake and frozen unstaked funds. Properties:

module Unstaked_frozen : sig ... end

Representation of Unstaked frozen deposits

module Unstaked_finalizable : sig ... end

Representation of unstaked finalizable tez

val tez_to_pseudo : round:[ `Down | `Up ] -> Tez.t -> Z.t -> Frozen_tez.t -> Z.t

Pseudotoken helpers

val pseudo_to_partial_tez : Z.t -> Z.t -> Frozen_tez.t -> Q.t

val stake_values_real : Tez.t -> Z.t -> Frozen_tez.t -> Z.t * Q.t

val unstake_values_real : Tez.t -> Z.t -> Frozen_tez.t -> Z.t * Q.t