Module Tez_staking_helpers.Frozen_tez

Frozen_tez represents frozen stake and frozen unstaked funds. Properties:

sum of all current partial tez is an integer

Can only add integer amounts

Can always subtract integer amount (if lower than frozen amount)

If subtracting partial amount, must be the whole frozen amount (for given contract). The remainder is then distributed equally amongst remaining accounts, to keep property 1.

All entries of current are positive, non zero.

type t = {

delegate : string;

initial : Tez.t;

self_current : Tez.t;

co_current : Partial_tez.t Tezos_base.TzPervasives.String.Map.t;

}

val pp : Stdlib.Format.formatter -> t -> unit

val init : Tez.t -> Tezos_base.TzPervasives.String.Map.key -> Tezos_base.TzPervasives.String.Map.key -> t

val union : t -> t -> t

val get : Tezos_base.TzPervasives.String.Map.key -> t -> Partial_tez.t

val total_co_current_q : Partial_tez.t Tezos_base.TzPervasives.String.Map.t -> Partial_tez.t

val total_co_current : t -> Tez.t

val total_current : t -> Tez.t

val total_current_with_limits : limit_of_staking_over_baking:Q.t -> t -> Tez.t

val add_q_to_all_co_current : Q.t -> Partial_tez.t Tezos_base.TzPervasives.String.Map.t -> Partial_tez.t Tezos_base.TzPervasives.String.Map.t

val add_tez_to_all_current : edge:Q.t -> limit:Q.t -> Tez.t -> t -> t

val sub_tez_from_all_current : limit:Q.t -> Tez.t -> t -> t * Tez.t * Tez.t

val add_current_q : Partial_tez.t -> Tezos_base.TzPervasives.String.Map.key -> t -> t

val add_current : Tez.t -> Tezos_base.TzPervasives.String.Map.key -> t -> t

val add_self_current : Tez.t -> t -> t

val add_init : Tez.t -> Tezos_base.TzPervasives.String.Map.key -> t -> t

val sub_current : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t -> Tezos_base.TzPervasives.String.Map.key -> t -> t * Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t

val sub_current_q : Q.t -> Tezos_base.TzPervasives.String.Map.key -> t -> t

val sub_current_and_init : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t -> Tezos_base.TzPervasives.String.Map.key -> t -> t * Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t

val slash : limit:Q.t -> Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Constants.Parametric.t -> Tez.t -> Tezos_protocol_021_PsQuebec.Protocol.Percentage.t -> t -> t * Tez.t * Tez.t