Module Tezos_021_PsQuebec_test_helpers.Tezt_helpers

Test registration

val register_test : __FILE__:string -> ?file_tags:string list -> title:string -> ?additional_tags:string list -> ?slow:bool -> (unit -> unit Lwt.t) -> unit

Registers a test. The protocol's name is added to the title and tags. File-specific title prefix and tags can also be specified.

val register_test_es : __FILE__:string -> ?file_tags:string list -> title:string -> ?additional_tags:string list -> ?slow:bool -> (unit -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t) -> unit

Same as register_test, but for a test function returning unit tzresult Lwt.t. If the result is an error, the test fails.