Module Tezos_baking_021_PsQuebec.Baking_lib

API

val bake : Tezos_client_021_PsQuebec.Protocol_client_context.full -> ?minimal_fees:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t -> ?minimal_nanotez_per_gas_unit:Q.t -> ?minimal_nanotez_per_byte:Q.t -> ?force_apply_from_round:int -> ?force:bool -> ?minimal_timestamp:bool -> ?extra_operations:Baking_configuration.Operations_source.t -> ?monitor_node_mempool:bool -> ?context_root_path:string -> ?dal_node_endpoint:Uri.t -> ?count:int -> ?votes:Baking_configuration.per_block_votes_config -> ?state_recorder:Baking_configuration.state_recorder_config -> Baking_state.consensus_key list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val preattest : Tezos_client_021_PsQuebec.Protocol_client_context.full -> ?force:bool -> Baking_state.consensus_key list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val attest : Tezos_client_021_PsQuebec.Protocol_client_context.full -> ?force:bool -> Baking_state.consensus_key list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val propose : Tezos_client_021_PsQuebec.Protocol_client_context.full -> ?minimal_fees:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Tez.t -> ?minimal_nanotez_per_gas_unit:Q.t -> ?minimal_nanotez_per_byte:Q.t -> ?force_apply_from_round:int -> ?force:bool -> ?minimal_timestamp:bool -> ?extra_operations:Baking_configuration.Operations_source.t -> ?context_root_path:string -> ?state_recorder:Baking_configuration.state_recorder_config -> Baking_state.consensus_key list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val repropose : Tezos_client_021_PsQuebec.Protocol_client_context.full -> ?force:bool -> ?force_round:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Round.t -> Baking_state.consensus_key list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

repropose tries to bake a new block proposal on the same level as the current head. If provided, the proposal will use the force_round argument as its reproposal round, otherwise the current tenderbake round will be used.