Module Tezos_baking_021_PsQuebec.Operation_worker

Launch processes to gather operations from the mempool and make them available for the baker.

Datatypes

type t

type candidate = {

hash : Tezos_base.TzPervasives.Block_hash.t;

round_watched : Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Round.t;

payload_hash_watched : Tezos_protocol_021_PsQuebec.Protocol.Block_payload_hash.t;

}

val candidate_encoding : candidate Tezos_base.TzPervasives.Data_encoding.t

type event =

| Prequorum_reached of candidate * Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Kind.preattestation Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.operation list

| Quorum_reached of candidate * Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Kind.attestation Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.operation list

Constructors

val create : ?monitor_node_operations:bool -> Tezos_client_021_PsQuebec.Protocol_client_context.full -> t Lwt.t

create ?monitor_node_operations cctxt creates a monitoring process to fetch operations for the baker to process.

parameter monitor_node_operations
monitor operations on the node (defaults: true). Set monitor_node_operations to false to only consider externally provided (non-node) operations.

Utilities

val retrieve_pending_operations : Tezos_client_021_PsQuebec.Protocol_client_context.full -> t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Accessors

val get_current_operations : t -> Operation_pool.pool

val get_quorum_event_stream : t -> event Lwt_stream.t

Observers

val monitor_preattestation_quorum : t -> consensus_threshold:int -> get_slot_voting_power: (slot:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Slot.t -> int option) -> candidate -> unit Lwt.t

val monitor_attestation_quorum : t -> consensus_threshold:int -> get_slot_voting_power: (slot:Tezos_protocol_021_PsQuebec.Protocol.Alpha_context.Slot.t -> int option) -> candidate -> unit Lwt.t

val cancel_monitoring : t -> unit

val shutdown_worker : t -> (unit, exn list) Stdlib.result Lwt.t